Add files via upload
keymap.c is my custom keymap that I prefer to use. There are issues with the Mousekeys only working partially.
rules.mk will be needed. It is a simple file, but it's used to enable VIA and Mousekeys.

readme.md and rev5_promicro_unibody_5x12_vial.json comes from the qmk_firmware respository that had Thumbsup in it.

thumbsup_rev5_promicro_unibody_5x12_default_5x12_v5_via.hex is the result of compiling the keymap.c with QMKSYS.